requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
pyaes==1.6.1
pyasn1==0.4.8
pyTelegramBotAPI==3.8.1
python-dotenv==0.18.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
Telethon==1.23.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
